# backend/public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Nurse Api",
        "description": "E-Nurse Api description",
        "contact": {
            "email": "developer@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Nurse login",
                "description": "Nurse login",
                "operationId": "login",
                "requestBody": {
                    "description": "Nurse login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "title": "Email",
                                        "type": "string",
                                        "example": "rami@example.com"
                                    },
                                    "password": {
                                        "title": "Password",
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "login succeeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "title": "Access Token",
                                            "type": "string",
                                            "example": "20d338931e8d6bd9466edeba7dfsdasg8365ev7wa75opasSsa5698ea7dce7c7bc01aa5cc5b4735691c50a2fe3228"
                                        },
                                        "token_type": {
                                            "title": "Token Type",
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "title": "Expire in",
                                            "type": "integer",
                                            "example": "3600"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register new Nurse",
                "description": "Register new Nurse",
                "operationId": "register",
                "requestBody": {
                    "description": "Register new Nurse",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Nurse"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registration succeeded"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Nurse logout",
                "description": "Nurse logout",
                "operationId": "logout",
                "responses": {
                    "201": {
                        "description": "successfully signed out"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh token",
                "description": "Refresh token",
                "operationId": "refresh",
                "responses": {
                    "201": {
                        "description": "Refresh succeeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "title": "Access Token",
                                            "type": "string",
                                            "example": "20d338931e8d6bd9466edeba7dfsdasg8365ev7wa75opasSsa5698ea7dce7c7bc01aa5cc5b4735691c50a2fe3228"
                                        },
                                        "token_type": {
                                            "title": "Token Type",
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "title": "Expire in",
                                            "type": "integer",
                                            "example": "3600"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/auth/nurseProfile": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Nurse Profile",
                "description": "Nurse Profile",
                "operationId": "nurseProfile",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Nurse": {
                                            "$ref": "#/components/schemas/Nurse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Nurse": {
                "title": "Nurse",
                "description": "Nurse",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Nurse id",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "Nurse Name (Max:50)",
                        "type": "string",
                        "example": "ahmad"
                    },
                    "email": {
                        "title": "email",
                        "description": "Nurse Email (Should be Unique)",
                        "type": "string",
                        "example": "ahmad@example.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "Nurse password (Min:8)",
                        "type": "string",
                        "example": "12345678"
                    },
                    "gender": {
                        "title": "gender",
                        "description": "Nurse gender",
                        "type": "boolean",
                        "example": "1"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "Nurse phone",
                        "type": "string",
                        "example": "0999111222"
                    },
                    "is_resigned": {
                        "title": "is_resigned",
                        "description": "Resigning status",
                        "type": "boolean",
                        "example": "0"
                    },
                    "is_admin": {
                        "title": "is_admin",
                        "description": "Nurse Role (admin|user)",
                        "type": "boolean",
                        "example": "0"
                    },
                    "created_at": {
                        "title": "Create Date",
                        "description": "Create Date)",
                        "type": "string",
                        "example": "2023-04-19 12:37:32 PM"
                    },
                    "updated_at": {
                        "title": "Update Date",
                        "description": "Update Date",
                        "type": "string",
                        "example": "2023-04-19 10:37:32 AM"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}